LayoutTests/media/modern-media-controls/mute-button/mute-button-expected.txt

Testing the MuteButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS muteButton.element.localName is "button"
PASS muteButton.element.classList.contains("mute") is true
PASS muteButton.muted is true
PASS muteButton.iconName became Icons.VolumeMutedRTL
PASS muteButton.image.element.style.webkitMaskImage.includes("macOS/VolumeMuted-RTL.svg") became true
Unmuting...
PASS muteButton.muted is false
PASS muteButton.iconName became Icons.Volume3RTL
PASS muteButton.image.element.style.webkitMaskImage.includes("macOS/Volume3-RTL.svg") became true
PASS successfullyParsed is true

TEST COMPLETE
